Add api plugins support and sdb plugin
Allow the optional addition of the sdb handler plugin for the
nymms API.

Also squash the task_context when saving results.  The task_context
can be large and causes errors when saving to SDB.
