+++ nymms/api/routes.py
@@ -11,7 +11,6 @@ from nymms.state import sdb_state
 from nymms.suppress import sdb_suppress
 from nymms.config import config
 from nymms import schemas
-from nymms.providers.sdb import SimpleDBBackend
 
 
 logger = logging.getLogger(__name__)
@@ -43,37 +42,6 @@ def state():
         filters=request.args,
         max_items=limit)
     return [s.to_primitive() for s in states]
-
-
-# Disabling for now- need a way of adding optional API endpoints
-# For example, this end point would only work if the SDB results handler
-# was enabled
-# @nymms_api.route('/result', methods=['GET'])
-def result():
-    """
-    List past results.
-
-    Query Params:
-    - limit (default 1000)
-    - from_timestamp
-    - to_timestamp
-    """
-    region = config.settings['region']
-    domain_name = config.settings['result_domain']
-    backend = SimpleDBBackend(region, domain_name)
-    args = request.args.to_dict(flat=True)
-    limit = int(args.pop('limit', DEFAULT_RESULT_LIMIT))
-    from_timestamp = args.pop('from_timestamp', None)
-    to_timestamp = args.pop('to_timestamp', None)
-    filters = []
-    if from_timestamp:
-        filters.append(
-            'timestamp >= "%s"' % arrow.get(from_timestamp))
-    if to_timestamp:
-        filters.append(
-            'timestamp <= "%s"' % arrow.get(to_timestamp).timestamp)
-    results, _ = backend.filter(filters=filters, max_items=limit)
-    return [schemas.APIResult(r).to_primitive() for r in results]
 
 
 @nymms_api.route('/suppress', methods=['GET', 'POST'])
